Attempt to get audit logs working
This is not working as wanted atm, just a push for the moment with a step in the right direction. I've changed some structs to better work with solidity, instead of storing a struct with data, which solidity has a hard thing to do when dealing with a struct that has strings I instead store the id of the version, which itself stores a version of the post to the user.